Small refactoring and moving files around
