+++ src-ext/services/quick-pick.ts
