+++ src-ext/services/task-manager.ts
